% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{privateschools}
\alias{privateschools}
\title{General information and location for 28,347 U.S. private schools}
\format{\preformatted{'data.frame':	28347 obs. of  45 variables:
 $ id               : int  1 2 3 4 5 6 7 8 9 10 ...
 $ biz_name         : Factor w/ 20816 levels "()EWING MENNONITE SCHOOL",..: 186 632 633 634 635 974 1029 1162 1568 1570 ...
 $ e_address        : Factor w/ 26993 levels "","05259 Church St",..: 26549 23156 18481 19015 17770 7465 25898 2776 13167 26516 ...
 $ e_city           : Factor w/ 6112 levels "6, Patuxent",..: 5220 105 105 105 105 105 105 105 1699 3352 ...
 $ e_state          : Factor w/ 51 levels "AK","AL","AR",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ e_postal         : int  99672 99518 99518 99507 99517 99515 99524 99516 99701 99926 ...
 $ e_zip_full       : Factor w/ 1540 levels "","35007-3035",..: 1519 1501 1500 1492 1499 1495 1502 1496 1521 1540 ...
 $ e_country        : Factor w/ 1 level "USA": 1 1 1 1 1 1 1 1 1 1 ...
 $ loc_county       : Factor w/ 1410 levels "","Abbeville",..: 662 29 29 29 29 29 29 29 435 1029 ...
 $ loc_area_code    : int  907 907 907 907 907 907 907 907 907 907 ...
 $ loc_FIPS         : int  2122 2020 2020 2020 2020 2020 2020 2020 2090 2198 ...
 $ loc_MSA          : int  NA 380 380 380 380 380 380 380 NA NA ...
 $ loc_PMSA         : int  NA NA NA NA NA NA NA NA NA NA ...
 $ loc_TZ           : Factor w/ 7 levels "","CST","EST",..: 6 6 6 6 6 6 6 6 6 6 ...
 $ loc_DST          : Factor w/ 3 levels "N","U","Y": 3 3 3 3 3 3 3 3 3 3 ...
 $ loc_LAT_centroid : num  60.5 61.2 61.2 61.2 61.2 ...
 $ loc_LAT_poly     : num  60.5 61.1 61.2 61.1 61.2 ...
 $ loc_LONG_centroid: num  -151 -150 -150 -150 -150 ...
 $ loc_LONG_poly    : num  -151 -150 -150 -150 -150 ...
 $ biz_phone        : Factor w/ 27391 levels "","(201) 209-9070",..: 24320 24046 24354 24333 24322 24327 24330 24331 24342 24366 ...
 $ grade_low        : int  3 2 2 3 2 11 2 3 3 7 ...
 $ grade_high       : int  17 13 4 15 11 17 13 16 17 15 ...
 $ school_days      : int  180 180 180 180 220 180 170 180 180 180 ...
 $ school_hours     : num  5 NA 5.5 7 10 6.5 6 4.5 7 6.5 ...
 $ f_lib            : Factor w/ 3 levels "N","U","Y": 3 3 3 3 3 1 1 1 3 3 ...
 $ enroll_ungraded  : int  NA NA NA NA NA NA NA NA NA NA ...
 $ enroll_prek      : int  NA 100 46 NA 156 NA 14 NA NA NA ...
 $ enroll_k         : int  2 17 15 5 17 NA 14 2 3 NA ...
 $ enroll_1         : int  1 14 NA 2 7 NA 8 2 3 NA ...
 $ enroll_2         : int  1 16 NA 3 8 NA 8 NA 7 3 ...
 $ enroll_3         : int  1 13 NA 6 3 NA 7 1 NA NA ...
 $ enroll_4         : int  1 19 NA NA 2 NA 6 2 8 2 ...
 $ enroll_5         : int  2 19 NA 6 4 NA 13 NA NA 1 ...
 $ enroll_6         : int  3 14 NA 2 3 2 10 3 5 NA ...
 $ enroll_7         : int  2 7 NA 3 NA 1 14 1 4 1 ...
 $ enroll_8         : int  2 11 NA 5 NA 3 8 NA NA NA ...
 $ enroll_9         : int  1 NA NA 1 NA 1 NA 2 6 NA ...
 $ enroll_10        : int  3 NA NA 1 NA 6 NA NA 4 1 ...
 $ enroll_11        : int  2 NA NA NA NA 6 NA 1 4 NA ...
 $ enroll_12        : int  1 NA NA NA NA 2 NA NA 1 NA ...
 $ enroll_total     : int  22 230 61 34 200 21 102 14 45 8 ...
 $ school_coed      : Factor w/ 4 levels "C","F","M","U": 1 1 1 1 1 1 1 1 1 1 ...
 $ school_relig     : Factor w/ 2 levels "N","Y": 2 2 2 2 2 2 2 2 2 2 ...
 $ school_religion  : Factor w/ 30 levels "","African Methodist Episcopal",..: 8 22 25 30 25 25 25 8 5 5 ...
 $ school_type      : Factor w/ 8 levels "","Alternative",..: 6 6 4 6 5 6 8 6 6 6 ...
}}
\usage{
privateschools
}
\description{
General information and location for 28,347 U.S. private schools
}
\keyword{datasets}

