% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publicschooldistricts}
\alias{publicschooldistricts}
\title{General information and location for 17,889 U.S. school districts}
\format{\preformatted{'data.frame':	17889 obs. of  28 variables:
 $ s_district_id    : int  200050 200010 200007 200180 200525 200020 200030 200730 200800 200070 ...
 $ s_state_id       : Factor w/ 14117 levels "01-000-0000-00",..: 5579 7068 9999 9111 10381 11175 12145 12831 246 1147 ...
 $ district_name    : Factor w/ 17423 levels "21ST CENTURY CHARTER SCH OF GARY",..: 225 264 265 456 495 1260 1781 2725 2898 3426 ...
 $ e_address        : Factor w/ 14226 levels "\\\\\\\\\\"\\\\\\"150 NORTH AVE",..: 12852 13092 12950 8917 13033 12851 12814 12741 11336 12736 ...
 $ e_city           : Factor w/ 8378 levels " BOX 261\\\\\\"\\\\\\"\\\\\\"\\"",..: 7449 161 6593 161 4699 7603 5077 173 161 2826 ...
 $ e_state          : Factor w/ 57 levels "AK","AL","AR",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ e_postal         : Factor w/ 13006 levels "10001","10002",..: 12989 12956 12972 12954 13004 12974 12968 12991 12955 12963 ...
 $ e_zip_full       : Factor w/ 15468 levels "","00820-4665",..: 1 1 15441 15430 15467 15443 15438 1 15431 1 ...
 $ e_country        : Factor w/ 4 levels "01227-0299","01756-0176",..: 4 4 4 4 4 4 4 4 4 4 ...
 $ loc_county       : Factor w/ 1835 levels "","Abbeville",..: 1544 37 18 37 1336 1173 204 773 37 1688 ...
 $ loc_area_code    : Factor w/ 248 levels "","201","202",..: 215 215 215 215 215 215 215 215 215 215 ...
 $ loc_FIPS         : int  2240 2020 2013 2020 2198 2180 2060 2105 2020 2261 ...
 $ loc_MSA          : int  NA 380 NA 380 NA NA NA NA 380 NA ...
 $ loc_PMSA         : int  NA NA NA NA NA NA NA NA NA NA ...
 $ loc_TZ           : Factor w/ 10 levels "","1120","CST",..: 8 8 8 8 8 8 8 8 8 8 ...
 $ loc_DST          : Factor w/ 4 levels "EST","N","U",..: 4 4 4 4 4 4 4 4 4 4 ...
 $ loc_LAT_centroid : Factor w/ 12448 levels "","13.4751","17.7473",..: 12437 12425 12399 12426 12398 12435 12415 12410 12424 12431 ...
 $ loc_LAT_poly     : num  63.3 61.2 55.3 61.2 55.1 ...
 $ loc_LONG_centroid: num  -142 -150 -161 -150 -131 ...
 $ loc_LONG_poly    : num  -143 -150 -160 -150 -132 ...
 $ cnt_students     : Factor w/ 5497 levels "","-70.6036395",..: 3741 3290 2359 3881 2336 1134 818 995 2061 3993 ...
 $ cnt_teachers     : int  31 6 34 2839 29 171 14 18 11 37 ...
 $ cnt_schools      : int  8 3 7 98 3 15 2 5 4 8 ...
 $ info_setting     : Factor w/ 14 levels "","0","City: Large",..: 8 8 8 3 8 8 8 8 3 8 ...
 $ ratio_teach      : Factor w/ 385 levels "","0","0.1","0.2",..: 71 337 361 96 374 379 22 369 157 62 ...
 $ dist_type        : Factor w/ 8 levels "","Component District",..: 6 6 6 6 6 6 6 6 6 6 ...
 $ biz_phone        : Factor w/ 16852 levels "","(201) 200-3200",..: 15004 14959 14963 14984 15005 14977 14956 14992 14970 14993 ...
 $ X.1              : Factor w/ 4 levels "","(413) 684-0320",..: 1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
publicschooldistricts
}
\description{
General information and location for 17,889 U.S. school districts
}
\keyword{datasets}

