% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{publicschools}
\alias{publicschools}
\title{General information and location for 101,385 U.S. public schools}
\format{\preformatted{'data.frame':	101385 obs. of  42 variables:
 $ X                : int  1 2 3 4 5 6 7 8 9 10 ...
 $ id               : int  1 2 3 4 5 6 7 8 9 10 ...
 $ biz_name         : Factor w/ 85726 levels "10TH GRADE ACADEMY - 76",..: 152 223 257 259 261 263 283 361 375 378 ...
 $ dist_name        : Factor w/ 16354 levels "21ST CENTURY CHARTER SCH OF GARY",..: 3512 4565 6332 6332 6332 6332 15921 1229 6734 6734 ...
 $ e_address        : Factor w/ 83093 levels "#1 HAWK DRIVE",..: 62566 54548 36513 826 48174 79917 79562 8127 35143 40498 ...
 $ e_city           : Factor w/ 14630 levels "4, Chaptico",..: 3096 538 3 3 3 3 116 1010 6100 6100 ...
 $ e_state          : Factor w/ 55 levels "AK","AL","AR",..: 2 2 2 2 2 2 2 2 2 2 ...
 $ e_postal         : int  36322 36502 36310 36310 36310 36310 36720 35020 35816 35810 ...
 $ e_zip_full       : Factor w/ 68498 levels "","00601-9701",..: 22621 22703 22615 22614 22612 22613 22880 21649 22301 22288 ...
 $ e_country        : Factor w/ 1 level "USA": 1 1 1 1 1 1 1 1 1 1 ...
 $ school_county    : Factor w/ 1956 levels "A?ASCO MUNICIPIO",..: 471 586 795 795 795 795 1894 881 1073 1073 ...
 $ loc_area_code    : int  334 251 334 334 334 334 334 205 256 256 ...
 $ loc_FIPS         : int  1045 1053 1067 1067 1067 1067 1131 1073 1089 1089 ...
 $ loc_MSA          : int  2180 NA NA NA NA NA NA 1000 3440 3440 ...
 $ loc_PMSA         : int  NA NA NA NA NA NA NA NA NA NA ...
 $ loc_TZ           : Factor w/ 10 levels "","CST","EST",..: 2 2 2 2 2 2 2 2 2 2 ...
 $ loc_DST          : Factor w/ 3 levels "N","U","Y": 3 3 3 3 3 3 3 3 3 3 ...
 $ loc_LAT_centroid : num  31.2 31.1 31.6 31.6 31.6 ...
 $ loc_LAT_poly     : num  31.3 31 31.6 31.6 31.6 ...
 $ loc_LONG_centroid: num  -85.8 -87.5 -85.3 -85.3 -85.3 ...
 $ loc_LONG_poly    : num  -85.7 -87.5 -85.3 -85.2 -85.3 ...
 $ biz_phone        : Factor w/ 92782 levels "","(201) 200-3200",..: 20554 8154 20546 20549 20547 20548 20544 1325 9500 9470 ...
 $ s_district_id    : int  101080 101350 101740 101740 101740 101740 103510 100330 101800 101800 ...
 $ grade_low        : Factor w/ 16 levels "0","1","10","11",..: 14 7 1 15 13 10 15 15 14 14 ...
 $ grade_high       : Factor w/ 16 levels "0","1","10","11",..: 8 8 1 9 5 12 10 10 12 12 ...
 $ loc_desc         : Factor w/ 13 levels "","City: Large",..: 6 11 5 5 5 5 7 8 3 3 ...
 $ flag_charter     : Factor w/ 3 levels "N","U","Y": 2 2 2 2 2 2 2 2 2 2 ...
 $ flag_magnet      : Factor w/ 3 levels "N","U","Y": 1 1 2 1 1 1 1 1 3 3 ...
 $ flag_title       : Factor w/ 3 levels "N","U","Y": 3 3 2 3 3 3 3 3 1 1 ...
 $ flag_title_all   : Factor w/ 3 levels "N","U","Y": 3 3 2 3 1 3 3 3 2 2 ...
 $ cnt_students     : int  475 287 NA 492 351 246 179 514 412 399 ...
 $ cnt_teachers     : num  32.2 17.9 0 33.7 26.3 14.3 14.2 34.8 33.6 33.7 ...
 $ teach_ratio      : num  14.8 16 NA 14.6 13.3 17.2 12.6 14.8 12.3 11.8 ...
 $ cnt_male         : int  240 143 NA 271 161 131 88 258 183 188 ...
 $ cnt_female       : int  235 144 NA 221 190 115 91 256 229 211 ...
 $ cnt_native       : int  4 0 NA 0 0 0 0 0 6 11 ...
 $ cnt_asian        : int  4 0 NA 2 5 3 0 0 4 14 ...
 $ cnt_black        : int  165 227 NA 280 205 168 179 505 226 233 ...
 $ cnt_hispanic     : int  55 3 NA 13 10 9 0 9 9 31 ...
 $ cnt_white        : int  247 57 NA 197 131 66 0 0 167 110 ...
 $ cnt_freelunch    : int  280 241 NA 368 232 185 161 477 83 154 ...
 $ cnt_reduced      : int  32 25 NA 40 33 21 0 16 28 49 ...
}}
\usage{
publicschools
}
\description{
General information and location for 101,385 U.S. public schools
}
\keyword{datasets}

